backend/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
groq==0.11.0
reportlab==4.2.5
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1  # Add this line
sqlalchemy==2.0.35
aiosqlite==0.20.0